+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [5.0.5] - 2021/06/27
+
+* MapOfSets.isEmptyForKey() and MapOfSets.isNotEmptyForKey().
+
 ## [5.0.4] - 2021/06/24
 
 * Introduce the Predicate typedef.